Extract root and tag parsers
